pyproject.toml
[tool.poetry]
name = "auth_service_tortoise"
version = "0.1.0"
description = ""
authors = ["Umesh Chaudhary <2umeshchaudhary@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.2"}
python-rapidjson = "^1.0"
ciso8601 = "^2.1.3"
pytest-asyncio = "^0.14.0"
httpx = "^0.17.1"
aerich = "^0.5.3"
pydantic = "^1.8.1"
orjson = "^3.5.1"

[tool.poetry.dev-dependencies]
tortoise-orm = {extras = ["sqlite"], version = "^0.17.2"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
